+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.1.0] - 2021-06-01
+### Changed
+- Natural Language Processing on timestamps using [Chrono](https://github.com/wanasit/chrono).  Introduction of --timeframe option for search.  Addition of developer logging via LogDNA for troubleshooting.
+- Changed CONTRIBUTING.md to reflect "fork-and-PR" process
+
+### TODO
+- Dev logs currently only in index.js.  Extend for full coverage
+- Add "today" when a date reference isn't explicity included (ie `search --from "7PM" --to "8PM"` instead of `search --from "today at 7PM" --to "today at 8PM"`)
+- Systematic testing
+
 ## [v2.0.0] - 2020-05-21
 ### Fixed
 - Fix `Cannot read property 'statusCode' of undefined` when updating [#40](https://github.com/logdna/logdna-cli/pull/40)